sanity check, monog_db option

+++ weave/pantry.py
@@ -20,7 +20,7 @@ class Pantry():
     """Facilitate user interaction with the index of a Weave data warehouse.
     """
 
-    def __init__(self, index: IndexABC, pantry_path="weave-test", **kwargs):
+    def __init__(self, index: IndexABC, pantry_path="weave-test", mongo_db=False, **kwargs):
         """Initialize Pantry object.
 
         A pantry is a collection of baskets. This class facilitates the upload,
@@ -75,6 +75,7 @@ class Pantry():
                            **kwargs
         )
         self.metadata['index_metadata'] = self.index.generate_metadata()
+        self.mongo_db = mongo_db
 
     def validate_path_in_pantry(self, path):
         """Validate the given path is within the pantry.
@@ -156,11 +157,12 @@ class Pantry():
                 "delete that basket before deleting its parent basket."
             )
 
-        MongoDB.remove_document(remove_item.iloc[0].uuid)
-
         self.validate_path_in_pantry(remove_item.iloc[0].address)
         self.index.untrack_basket(remove_item.iloc[0].address, **kwargs)
         self.file_system.rm(remove_item.iloc[0].address, recursive=True)
+
+        if self.mongo_db:
+            MongoDB.remove_document(remove_item.iloc[0].uuid)
 
     def upload_basket(self, upload_items, basket_type, **kwargs):
         """Upload a basket to the same pantry referenced by the Index
@@ -215,7 +217,9 @@ class Pantry():
 
         single_indice_index = create_index_from_fs(up_dir, self.file_system)
         self.index.track_basket(single_indice_index)
-        MongoDB.append_document(single_indice_index, self)
+
+        if self.mongo_db:
+            MongoDB.append_document(single_indice_index, self)
 
         return single_indice_index
